Move hooks (SpecialPageList, ApiModuleManager)
