+++ src/Setup.php
@@ -23,98 +23,7 @@ final class Setup {
 	 * @since 3.0
 	 */
 	public static function initExtension( &$vars ) {
-
-		/**
-		 * @see https://www.mediawiki.org/wiki/Localisation#Localising_namespaces_and_special_page_aliases
-		 */
-		$vars['wgMessagesDirs']['SemanticMediaWiki'] = $vars['smwgIP'] . 'i18n';
-		$vars['wgExtensionMessagesFiles']['SemanticMediaWikiAlias'] = $vars['smwgIP'] . 'i18n/extra/SemanticMediaWiki.alias.php';
-		$vars['wgExtensionMessagesFiles']['SemanticMediaWikiMagic'] = $vars['smwgIP'] . 'i18n/extra/SemanticMediaWiki.magic.php';
-
 		Hooks::registerEarly( $vars );
-	}
-
-	/**
-	 * @see Hooks::initExtension
-	 */
-	public static function getAPIModules() {
-
-		if ( !ApplicationFactory::getInstance()->getSettings()->get( 'smwgSemanticsEnabled' ) ) {
-			return [];
-		}
-
-		return [
-			'smwinfo' => '\SMW\MediaWiki\Api\Info',
-			'smwtask' => '\SMW\MediaWiki\Api\Task',
-			'smwbrowse' => '\SMW\MediaWiki\Api\Browse',
-			'ask' => '\SMW\MediaWiki\Api\Ask',
-			'askargs' => '\SMW\MediaWiki\Api\AskArgs',
-			'browsebysubject' => '\SMW\MediaWiki\Api\BrowseBySubject',
-			'browsebyproperty' => '\SMW\MediaWiki\Api\BrowseByProperty'
-		];
-	}
-
-	/**
-	 * @see Hooks::initExtension
-	 */
-	public static function initSpecialPageList( array &$specialPages ) {
-
-		if ( !ApplicationFactory::getInstance()->getSettings()->get( 'smwgSemanticsEnabled' ) ) {
-			return;
-		}
-
-		$specials = [
-			'ExportRDF' => [
-				'page' => 'SMWSpecialOWLExport'
-			],
-			'SMWAdmin' => [
-				'page' => 'SMW\MediaWiki\Specials\SpecialAdmin'
-			],
-			'Ask' => [
-				'page' => 'SMW\MediaWiki\Specials\SpecialAsk'
-			],
-			'Browse' => [
-				'page' =>  'SMW\MediaWiki\Specials\SpecialBrowse'
-			],
-			'Concepts' => [
-				'page' => 'SMW\SpecialConcepts'
-			],
-			'PageProperty' => [
-				'page' =>  'SMW\MediaWiki\Specials\SpecialPageProperty'
-			],
-			'SearchByProperty' => [
-				'page' => 'SMW\MediaWiki\Specials\SpecialSearchByProperty'
-			],
-			'PropertyLabelSimilarity' => [
-				'page' => 'SMW\MediaWiki\Specials\SpecialPropertyLabelSimilarity'
-			],
-			'ProcessingErrorList' => [
-				'page' => 'SMW\MediaWiki\Specials\SpecialProcessingErrorList'
-			],
-			'MissingRedirectAnnotations' => [
-				'page' => 'SMW\MediaWiki\Specials\SpecialMissingRedirectAnnotations'
-			],
-			'Types' => [
-				'page' => 'SMWSpecialTypes'
-			],
-			'URIResolver' => [
-				'page' => 'SMW\MediaWiki\Specials\SpecialURIResolver'
-			],
-			'Properties' => [
-				'page' => 'SMW\SpecialProperties'
-			],
-			'UnusedProperties' => [
-				'page' => 'SMW\SpecialUnusedProperties'
-			],
-			'WantedProperties' => [
-				'page' => 'SMW\SpecialWantedProperties'
-			]
-		];
-
-		// Register data
-		foreach ( $specials as $special => $page ) {
-			$specialPages[$special] = $page['page'];
-		}
 	}
 
 	/**
